+++ src/components/Admin/Admin.jsx
@@ -1,13 +1,35 @@
 import axios from 'axios';
-import {useSelector} from 'react-redux';
+import {useSelector, useDispatch} from 'react-redux';
 import { useEffect, useState } from 'react';
 
 
 function Admin(){
 
+    const dispatch=useDispatch();
+
+    useEffect(() => {
+        fetchOrders();
+      }, []);
+
+    const fetchOrders=()=>{
+        axios({
+            method: 'GET',
+            url: '/api/order'
+        })
+            .then((response)=>{
+                let orders=response.data;
+                console.log('orders is', orders);
+                dispatch({
+                  type:'SET_ORDERS',
+                  payload: orders,
+                })
+            })
+            .catch(err=>console.error('in get orders api err', err));
+    };
+
     const orders = useSelector(store => store.adminOrders)
 
-
+    console.log('orders is: in function not get:', orders)
     return(
         <table className='table'>
             <thead>
